Fix type from int to float for mag_axis_r and mag_axis_z.

+++ l2g/equil/_equilibrium.py
@@ -68,8 +68,8 @@ class Equilibrium(object):
         self.wall_contour_r: list = []
         self.wall_contour_z: list = []
 
-        self.mag_axis_r: int = -1
-        self.mag_axis_z: int = -1
+        self.mag_axis_r: float = -1.0
+        self.mag_axis_z: float = -1.0
 
         self.grid_dim_r: int = -1
         self.grid_dim_z: int = -1
